refactor fenced JSON extraction
